+++ management/server/core/hub-manager/hub-manager-impl/src/main/java/io/subutai/core/hubmanager/impl/model/RhP2PIpEntity.java
@@ -1,4 +1,4 @@
-package io.subutai.core.hubmanager.impl.entity;
+package io.subutai.core.hubmanager.impl.model;
 
 
 import javax.persistence.Column;
