+++ management/server/core/hub-manager/hub-manager-impl/src/main/java/io/subutai/core/hubmanager/impl/processor/HubEnvironmentProcessor.java
@@ -1,4 +1,4 @@
-package io.subutai.core.hubmanager.impl.environment;
+package io.subutai.core.hubmanager.impl.processor;
 
 
 import java.util.Set;
